Refactor benchmark
